+++ src/main/java/com/example/ctaegorymanagment/EndPoints/SubCategoryRestController.java
@@ -25,31 +25,32 @@ public class SubCategoryRestController {
 
     @GetMapping("/getAllSubCategories")
     public List<Subcategory> getAllSubcategories() {
+
         return subCategoryService.getAllSubcategories();
     }
 
     @GetMapping("/getById/{id}")
-    public ResponseEntity<Subcategory> getSubcategoryById(@PathVariable Long id) {
+    public ResponseEntity<Subcategory> getSubcategoryById(@PathVariable int id) {
         return   subCategoryService.getSubcategoryById(id)
                 .map(subcategory -> new ResponseEntity<>(subcategory, HttpStatus.OK))
                 .orElse(new ResponseEntity<>(HttpStatus.NOT_FOUND));
     }
 
     @PutMapping("/update/{id}")
-    public ResponseEntity<Subcategory> updateSubcategory(@PathVariable Long id, @RequestBody Subcategory updatedSubcategory) {
+    public ResponseEntity<Subcategory> updateSubcategory(@PathVariable int id, @RequestBody Subcategory updatedSubcategory) {
         Subcategory subcategory = subCategoryService.updateSubcategory(id, updatedSubcategory);
         return (subcategory != null) ? ResponseEntity.ok(subcategory) : ResponseEntity.notFound().build();
     }
 
     @DeleteMapping("delete/{id}")
-    public ResponseEntity<Void> deleteSubcategory(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteSubcategory(@PathVariable int id) {
         boolean deleted = subCategoryService.deleteSubcategory(id);
         return (deleted) ? ResponseEntity.noContent().build() : ResponseEntity.notFound().build();
     }
 
 
     @GetMapping("/byCategory/{categoryId}")
-    public ResponseEntity<List<Subcategory>> getSubcategoriesByCategoryId(@PathVariable Long categoryId) {
+    public ResponseEntity<List<Subcategory>> getSubcategoriesByCategoryId(@PathVariable int categoryId) {
         try {
             List<Subcategory> subcategories = subCategoryService.getSubcategoriesByCategoryId(categoryId);
             return ResponseEntity.ok(subcategories);
